+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Current
 
+- Add: `DEXPair.get_link()` to get a direct link to the trading pair page
 - Add: `DEXPair.is_tradeable()` for quick checks if a pair has decent volume
 - Add: `PandasPairUniverse.get_exchange_for_pair`
 - Add: Aave Ghost (GHST) stablecoin
 - Add: Jarvis Synthetic Euro (jEUR) stablecoin
 - API update: `get_pair_by_human_description()` can have exchange slug set to `None`
   matching the best trading pair across with the lowest fee across all DEXes
+- API update: `get_closest_pair(pair)` can now now take `DEXPair` instead of a primary key
+  as a lookup parameter and also display the pair name in error messages
 
 # 0.20.7